Refactor GameLoopManager and Add combat speed control x1/x2/x4 #554 #552 #550
