+++ Lords_Frontiers/Content/Game/Assets/UI/GameHUD/Texture/UpperHUD/buttonSkip.uasset
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:aa3dc166fd54d62e832c5dbea7009e24bea072b723566b2b9fde33a29500303e
+size 39235
