+++ Lords_Frontiers/Content/Game/Assets/UI/GameHUD/WBP_GameHUD.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:a8e96fe83351d0e937c7756f4e334f7a6013620c1027d8d8757dcf1e3ab1357e
-size 173342
+oid sha256:aba0a45978b4c6b18585f9cad97a92b6e21096fe924b2865dd9cb8b5d4d42933
+size 180124
