+++ Lords_Frontiers/Content/Game/Assets/UI/Tooltip/Config/DA_BuildingsUI.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:914827daeeda993e0ae7aced9ebf2c2f3cec79eff936377967856e0321cb980e
-size 15444
+oid sha256:dbe0094d28f0b99ce1668bf8a5995510de85f52efe35e6a10f30208b48e221d0
+size 10736
