+++ Lords_Frontiers/Content/Game/Assets/UI/Tooltip/W_Tooltip_Economy.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:b4b5f899ac95cbd054d97eecfa576049e8381faeec8f19be2882c495c9f1e847
-size 71149
+oid sha256:c0ea6e77cf7c2975dfbc01e7564a35206f5c9f52ba901f45fcf466d265bb98cb
+size 52194
