+++ Lords_Frontiers/Source/Lords_Frontiers/Private/Core/GameLoop/GameLoopConfig.cpp
@@ -1,4 +1,4 @@
-#include "Core/GameLoopConfig.h"
+#include "Core/GameLoop/GameLoopManager.h"
 
 UGameLoopConfig::UGameLoopConfig()
 {
