+++ Lords_Frontiers/Source/Lords_Frontiers/Public/UI/GameHUD.h
@@ -1,5 +1,6 @@
 #pragma once
-#include "Core/GameLoopManager.h"
+#include "Core/GameLoop/GameLoopManager.h"
+#include "Core/GameSessionController.h"
 #include "Lords_Frontiers/Public/UI/Widgets/ResourceItemWidget.h"
 #include "Lords_Frontiers/Public/Units/Unit.h"
 #include "Lords_Frontiers/Public/Waves/EnemyGroupSpawnPoint.h"
@@ -124,6 +125,12 @@ public:
 
 	UPROPERTY( meta = ( BindWidget ) )
 	UTextBlock* TextTimer;
+
+	UPROPERTY( meta = ( BindWidgetOptional ) )
+	TObjectPtr<UButton> ButtonSpeed;
+
+	UPROPERTY( meta = ( BindWidgetOptional ) )
+	TObjectPtr<UTextBlock> TextSpeed;
 
 	UPROPERTY( meta = ( BindWidgetOptional ) )
 	TObjectPtr<UTextBlock> Text_GoldIncome;
@@ -463,5 +470,13 @@ protected:
 	void OnWaveInfoButtonClicked();
 
 	UFUNCTION()
-	void HandleGameEnded( bool bVictory );
+	void HandleGameEnded( EGameResult Result );
+
+	UFUNCTION()
+	void OnSpeedButtonClicked();
+
+	UFUNCTION()
+	void HandleSpeedChanged( float NewSpeed );
+
+	void UpdateSpeedButtonVisibility( EGameLoopPhase Phase );
 };
